Useful function and a bug
Fixed the bug were when the x axis is greater than 1000, there is an error. This is due to the fact that once the value of x goes above 1000, the canvasIndex becomes 1 and 1000 is subtracted from the x axis. For example, if we were to place a pixel at 100,1200, the coordinate part of the request would be coordinate : { x : 1000 , y : 200} and the canvasIndex will be 1. 

Additionally, I have added a function in the Place class that allows you to see which user has placed that certain pixel at that single location. I also have added a little bit of handling when the coordinate of y is greater than 1000 since that is not possible. When this occurs, the class will raise an error, telling the user that the y coordinate can not be greater than 1000.